+++ src/css/snake.css
@@ -17,7 +17,9 @@ body {
 }
 
 .snake-start {
+  cursor: pointer;
   display: block;
+  border: 1px solid gold;
   width: 200px;
   margin: 0 auto;
   background-color: transparent;
@@ -25,4 +27,8 @@ body {
   font-size: 25px;
   text-align: center;
   text-transform: uppercase;
+}
+
+.snake-start:disabled {
+  cursor: not-allowed;
 }
