+++ src/css/styles.css
@@ -214,32 +214,3 @@ input[type='radio'].visually-hidden {
   top: 0;
   left: 0;
 }
-
-body {
-  /* background-color: black; */
-}
-
-.snake-title {
-  color: gold;
-  text-align: center;
-}
-
-.canvas-wrp {
-  /* height: 500px;
-  width: 500px; */
-  display: flex;
-  justify-content: center;
-  margin: auto;
-  margin-bottom: 36px;
-}
-
-.snake-start {
-  display: block;
-  width: 200px;
-  margin: 0 auto;
-  background-color: transparent;
-  color: gold;
-  font-size: 25px;
-  text-align: center;
-  text-transform: uppercase;
-}
