+++ src/js/scrypt.js
@@ -1,2 +1,2 @@
-// import Refs from './refs';
-// import './firebase';
+import Refs from './refs';
+import './firebase';
