+++ src/js/snake/Block.js
@@ -1,0 +1,29 @@
+import circle from './circle';
+import { blockSize } from './blockSizes';
+import { ctx } from './snake';
+class Block {
+  constructor(col, row) {
+    this.col = col;
+    this.row = row;
+  }
+
+  drawSquare(color) {
+    const x = this.col * blockSize;
+    const y = this.row * blockSize;
+    ctx.fillStyle = color;
+    ctx.fillRect(x, y, blockSize, blockSize);
+  }
+
+  drawCircle(color) {
+    const centerX = this.col * blockSize + blockSize / 2;
+    const centerY = this.row * blockSize + blockSize / 2;
+    ctx.fillStyle = color;
+    circle(centerX, centerY, blockSize / 2, true);
+  }
+
+  equal(otherBlock) {
+    return this.col === otherBlock.col && this.row === otherBlock.row;
+  }
+}
+
+export default Block;
