+++ src/js/snake/blockSizes.js
@@ -1,0 +1,9 @@
+import { canvas } from './snakeRefs';
+
+export const width = canvas.width;
+export const height = canvas.height;
+
+//=======blocks======
+export const blockSize = 10;
+export const widthInBlocks = width / blockSize;
+export const heightInBlocks = height / blockSize;
