+++ src/js/snake/circle.js
@@ -1,0 +1,13 @@
+import { ctx } from './snake';
+
+//========circle========
+
+const circle = function (x, y, radius, fillCircle) {
+  ctx.beginPath();
+  ctx.arc(x, y, radius, 0, Math.PI * 2, false);
+  if (fillCircle) {
+    ctx.fill();
+  }
+};
+
+export default circle;
