+++ src/js/snake/directions.js
@@ -1,0 +1,12 @@
+const directions = {
+  ArrowLeft: 'left',
+  ArrowUp: 'up',
+  ArrowRight: 'right',
+  ArrowDown: 'down',
+  KeyA: 'left',
+  KeyW: 'up',
+  KeyD: 'right',
+  KeyS: 'down',
+};
+
+export default directions;
