+++ src/js/snake/modes.js
@@ -1,0 +1,2 @@
+export const MODE_CLASSIC = 'classic';
+export const MODE_IGNORE_WALLS_COLLISION = 'ignoreWallsCollisionMode';
