+++ src/js/snake/snakeRefs.js
@@ -1,0 +1,3 @@
+export const startButton = document.querySelector('.snake-start');
+export const canvas = document.querySelector('#canvas');
+export const modeWrp = document.querySelector('.mode-wrp');
